- Fixes invalid anchorEl warning -- #207 - Using a drawer to show the menu (and sub-menus) on mobile screens.
