+++ src/client/components/defaults.js
@@ -1,6 +1,7 @@
 export const DEFAULT_PADDING = 10;
 export const HEADER_HEIGHT = 50;
 export const LEFT_DRAWER_WIDTH = 320;
+export const RIGHT_DRAWER_WIDTH = 320;
 /** The height of the bottom drawer when collapsed */
 export const BOTTOM_DRAWER_HEIGHT = 48;
 /** Whether the bottom drawer must be open (expanded) by default */
