+++ src/client/components/network-editor/controller.js
@@ -1,14 +1,22 @@
+import React from 'react';
 import EventEmitter from 'eventemitter3';
 import Cytoscape from 'cytoscape'; // eslint-disable-line
 import _ from 'lodash';
 import ReactDOMServer from 'react-dom/server';
+import JSZip from 'jszip';
+import { Canvg, presets } from 'canvg';
+import { saveAs } from 'file-saver';
 
 import { DEFAULT_PADDING } from '../defaults';
 import { clusterColor } from './network-style';
 import { monkeyPatchMathRandom, restoreMathRandom } from '../../rng';
 import { SearchController } from './search-contoller';
 import { UndoHandler } from './undo-stack';
-import { ExpandIcon, CollapseIcon } from './share-panel';
+import { getLegendSVG } from './legend-svg';
+import { stringToBlob } from './util';
+
+import ZoomOutMapIcon from '@material-ui/icons/ZoomOutMap';
+import { ZoomInIcon } from '../svg-icons';
 
 // Clusters that have this many nodes get optimized.
 // Note we are using number of nodes as a proxy for number of edges, assuming large clusters are mostly complete.
@@ -23,6 +31,12 @@ export const CoSELayoutOptions = {
   randomize: true,
   animate: false,
   padding: DEFAULT_PADDING,
+};
+
+export const ImageSize = {
+  SMALL:  { value:'SMALL',  scale: 0.5 },
+  MEDIUM: { value:'MEDIUM', scale: 1.0 },
+  LARGE:  { value:'LARGE',  scale: 2.0 },
 };
 
 /**
@@ -377,7 +391,7 @@ export class NetworkEditorController {
 
     const setButtonHTML = (elem, node) => {
       const collapsed = node.data('collapsed');
-      const jsx = collapsed ? ExpandIcon() : CollapseIcon();
+      const jsx = collapsed ? <ZoomOutMapIcon /> : <ZoomInIcon />;
       const html = ReactDOMServer.renderToStaticMarkup(jsx);
       elem.innerHTML = html;
     };
@@ -580,5 +594,129 @@ export class NetworkEditorController {
 
       return geneSet;
     }
+  }
+  
+
+  async exportImageArchive() {
+    const blobs = await Promise.all([
+      this.createNetworkImageBlob(ImageSize.SMALL),
+      this.createNetworkImageBlob(ImageSize.MEDIUM),
+      this.createNetworkImageBlob(ImageSize.LARGE),
+      this.createSVGLegendBlob()
+    ]);
+  
+    const zip = new JSZip();
+    zip.file('enrichment_map_small.png',  blobs[0]);
+    zip.file('enrichment_map_medium.png', blobs[1]);
+    zip.file('enrichment_map_large.png',  blobs[2]);
+    zip.file('node_color_legend_NES.svg', blobs[3]);
+  
+    this.saveZip(zip, 'images');
+  }
+
+  async exportDataArchive() {
+    const netID = this.networkIDStr;
+  
+    const fetchExport = async path => {
+      const res = await fetch(path);
+      return await res.text();
+    };
+  
+    const files = await Promise.all([
+      fetchExport(`/api/export/enrichment/${netID}`),
+      fetchExport(`/api/export/ranks/${netID}`),
+      fetchExport(`/api/export/gmt/${netID}`),
+    ]);
+  
+    const zip = new JSZip();
+    zip.file('enrichment_results.txt', files[0]);
+    zip.file('ranks.txt', files[1]);
+    zip.file('gene_sets.gmt', files[2]);
+  
+    this.saveZip(zip, 'enrichment');
+  }
+
+  async saveGeneList(genesJSON, pathways) { // used by the gene list panel (actually left-drawer.js)
+    const lines = ['gene\trank'];
+    for(const { gene, rank } of genesJSON) {
+      lines.push(`${gene}\t${rank}`);
+    }
+    const fullText = lines.join('\n');
+    const blob = stringToBlob(fullText);
+  
+    let fileName = 'gene_ranks.txt';
+    if(pathways && pathways.length == 1) {
+      fileName = `gene_ranks_(${pathways[0]}).txt`;
+    } else if(pathways && pathways.length > 1) {
+      fileName = `gene_ranks_${pathways.length}_pathways.txt`;
+    }
+  
+    saveAs(blob, fileName);
+  }
+
+  async createNetworkImageBlob(imageSize) {
+    const { cy, bubbleSets } = this;
+    const renderer = cy.renderer();
+  
+    // render the network to a buffer canvas
+    const cyoptions = {
+      output: 'blob',
+      bg: 'white',
+      full: true, // full must be true for the calculations below to work
+      scale: imageSize.scale,
+    };
+    const cyCanvas = renderer.bufferCanvasImage(cyoptions);
+    const { width, height } = cyCanvas;
+  
+    // compute the transform to be applied to the bubbleSet svg layer
+    // this code was adapted from the code in renderer.bufferCanvasImage()
+    var bb = cy.elements().boundingBox();
+    const pxRatio = renderer.getPixelRatio();
+    const scale = imageSize.scale * pxRatio;
+    const dx = -bb.x1 * scale;
+    const dy = -bb.y1 * scale;
+    const transform = `translate(${dx},${dy})scale(${scale})`;
+  
+    // get the bubbleSet svg element
+    const svgElem = bubbleSets.layer.node.parentNode.cloneNode(true);
+    svgElem.firstChild.setAttribute('transform', transform); // firstChild is a <g> tag
+  
+    // render the bubbleSet svg layer using Canvg library
+    const svgCanvas = new OffscreenCanvas(width, height);
+    const ctx = svgCanvas.getContext('2d');
+    const svgRenderer = await Canvg.from(ctx, svgElem.innerHTML, presets.offscreen());
+    await svgRenderer.render();
+  
+    // combine the layers
+    const combinedCanvas = new OffscreenCanvas(width, height);
+    const combinedCtx = combinedCanvas.getContext('2d');
+    combinedCtx.drawImage(cyCanvas,  0, 0);
+    combinedCtx.drawImage(svgCanvas, 0, 0);
+  
+    const blob = await combinedCanvas.convertToBlob();
+    return blob;
+  }
+
+  async createSVGLegendBlob() {
+    const svg = getLegendSVG(this);
+    return new Blob([svg], { type: 'text/plain;charset=utf-8' });
+  }
+   
+  getZipFileName(suffix) {
+    const networkName = this.cy.data('name');
+    if(networkName) {
+      // eslint-disable-next-line no-control-regex
+      const reserved = /[<>:"/\\|?*\u0000-\u001F]/g;
+      if(!reserved.test(networkName)) {
+        return `${networkName}_${suffix}.zip`;
+      }
+    }
+    return `enrichment_map_${suffix}.zip`;
+  }
+
+  async saveZip(zip, type) {
+    const archiveBlob = await zip.generateAsync({ type: 'blob' });
+    const fileName = this.getZipFileName(type);
+    await saveAs(archiveBlob, fileName);
   }
 }
