+++ src/client/components/network-editor/main.js
@@ -1,14 +1,39 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
+import Mousetrap from 'mousetrap';
+import chroma from 'chroma-js';
 
 import { makeStyles } from '@material-ui/core/styles';
 
-import { HEADER_HEIGHT, LEFT_DRAWER_WIDTH, BOTTOM_DRAWER_HEIGHT, bottomDrawerHeight } from '../defaults';
+import { DEFAULT_PADDING, HEADER_HEIGHT, LEFT_DRAWER_WIDTH, BOTTOM_DRAWER_HEIGHT, bottomDrawerHeight } from '../defaults';
 import { EventEmitterProxy } from '../../../model/event-emitter-proxy';
 import { NetworkEditorController } from './controller';
+import { Header } from './header';
 import LeftDrawer from './left-drawer';
+import RightDrawer from './right-drawer';
 import BottomDrawer from './bottom-drawer';
+import PopoverMenu from './popover-menu';
+import { TYPE as UNDO_TYPE } from './undo-stack';
+import { delay } from './util';
+
+import { Button, IconButton, Dialog, DialogActions, DialogContent, DialogTitle } from '@material-ui/core';
+import { Snackbar, SnackbarContent } from '@material-ui/core';
+
+import CloseIcon from '@material-ui/icons/Close';
+import CircularProgressIcon from '@material-ui/core/CircularProgress';
+import FitScreenIcon from '@material-ui/icons/SettingsOverscan';
+import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
+import AddIcon from '@material-ui/icons/Add';
+import RemoveIcon from '@material-ui/icons/Remove';
+import UndoIcon from '@material-ui/icons/Undo';
+import DeleteIcon from '@material-ui/icons/Delete';
+import RestoreIcon from '@material-ui/icons/SettingsBackupRestore';
+import LinkIcon from '@material-ui/icons/Link';
+import InsertDriveFileOutlinedIcon from '@material-ui/icons/InsertDriveFileOutlined';
+
+
+const SHARE_MENU_ID  = "menu-share";
 
 
 const useStyles = makeStyles((theme) => ({
@@ -48,7 +73,17 @@ const useStyles = makeStyles((theme) => ({
       duration: theme.transitions.duration.enteringScreen,
     }),
   },
+  snackBar: {
+    top: '70px',
+    zOrder: 1000,
+  },
+  snackBarContent: {
+    color: 'inherit',
+    background: chroma(theme.palette.background.default).alpha(0.8).hex(),
+    backdropFilter: 'blur(6px)',
+  },
 }));
+
 
 const NetworkBackground = ({ controller }) => {
   const [ bgColor, setBgColor ] = useState('white');
@@ -67,59 +102,362 @@ const NetworkBackground = ({ controller }) => {
     <div style={{ position: 'absolute', zIndex: -1, width: '100%', height: '100%', backgroundColor: bgColor }} />
   );
 };
+NetworkBackground.propTypes = {
+  controller: PropTypes.instanceOf(NetworkEditorController).isRequired,
+};
+
+function RestoreConfirmDialog({ open, onOk, onCancel }) {
+  return (
+    <Dialog maxWidth="xs" open={open}>
+      <DialogTitle>Confirm Restore Network Layout</DialogTitle>
+      <DialogContent dividers>
+        <p>Are you sure you want to restore the network layout to its initial state?</p>
+        <p>All nodes will be returned to their initial positions. All deleted nodes will be restored.</p>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onCancel} autoFocus>Cancel</Button>
+        <Button onClick={onOk}>Ok</Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
+RestoreConfirmDialog.propTypes = {
+  open: PropTypes.bool.isRequired,
+  onOk: PropTypes.func.isRequired,
+  onCancel: PropTypes.func.isRequired
+};
+
+
+function getUndoMenuTitle(undoType) {
+  switch(undoType) {
+    case UNDO_TYPE.DELETE: return "Undo Delete";
+    case UNDO_TYPE.POSITION: return "Undo Move";
+    default: return "Undo";
+  }
+}
+
+function createPanner({ cy }) {
+  const zoomButtonFactor = 1.5;
+  const panFactor = 100;
+  const easing = 'ease-out';
+  const duration = 400;
+
+  const zoomByFactor = (factor) => {
+    const container = cy.container();
+    const x = container.clientWidth / 2;
+    const y = container.clientHeight / 2;
+    const zoom = cy.zoom() * factor;
+    cy.stop().animate({
+      zoom: {
+        level: zoom,
+        renderedPosition: { x, y }
+      },
+      easing,
+      duration,
+    });
+  };
+
+  const panBy = (x, y) => {
+    const z = 1;
+    cy.stop().animate({
+      panBy: { x: x * z, y: y * z },
+      easing,
+      duration,
+    });
+  };
+
+  const fit = () => {
+    cy.stop().animate({
+      fit: { padding: DEFAULT_PADDING },
+      easing,
+      duration,
+    });
+  };
+
+  return {
+    panLeft:  () => panBy(panFactor, 0),
+    panRight: () => panBy(-panFactor, 0),
+    panUp:    () => panBy(0, panFactor),
+    panDown:  () => panBy(0, -panFactor),
+    zoomOut:  () => zoomByFactor(1 / zoomButtonFactor),
+    zoomIn:   () => zoomByFactor(zoomButtonFactor),
+    fit:      () => fit(),
+  };
+}
+
+function handleCopyToClipboard() {
+  const url = window.location.href;
+  navigator.clipboard.writeText(url);
+}
+
+function snackBarOps(setSnackBarState) {
+  return {
+    close: () => setSnackBarState({ open: false }),
+    showMessage: message => setSnackBarState({ open: true, closeable: true, autoHideDelay: 3000, message }),
+    showSpinner: message => setSnackBarState({ open: true, closeable: false, spinner: true, message }),
+  };
+}
 
 const Main = ({
   controller,
   openLeftDrawer,
+  openRightDrawer,
   openBottomDrawer,
   isMobile,
   onContentClick,
   onCloseLeftDrawer,
+  onCloseRightDrawer,
+  onOpenLeftDrawer,
+  onOpenRightDrawer,
   onToggleBottomDrawer
 }) => {
+  const [ menuName, setMenuName ] = useState(null);
+  const [ anchorEl, setAnchorEl ] = useState(null);
+  const [ confirmDialogOpen, setConfirmDialogOpen ] = useState(false);
+  const [ undoEnabled, setUndoEnabled ] = useState(false);
+  const [ undoType, setUndoType] = useState(null);
+  const [ panner ] = useState(() => createPanner(controller));
+  const [ imageExportEnabled, setImageExportEnabled ] = useState(true);
+  const [ dataExportEnabled, setDataExportEnabled ] = useState(true);
+  const [ snackBarState, setSnackBarState ] = useState({
+    open: false,
+    message: "",
+    autoHideDelay: 4000,
+    closeable: true,
+    spinner: false
+  });
+
   const classes = useStyles();
+
+  const snack = snackBarOps(setSnackBarState);
 
   const shiftXCy = openLeftDrawer && !isMobile;
   const shiftYCy = openBottomDrawer;
 
+  const showShareMenu = (event) => {
+    setMenuName(SHARE_MENU_ID);
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleMenuClose = () => {
+    setMenuName(null);
+    setAnchorEl(null);
+  };
+
+  const handleNetworkRestore = () => {
+    setConfirmDialogOpen(true);
+  };
+  const onConfirmCancel = () => {
+    setConfirmDialogOpen(false);
+  };
+  const onConfirmOk = () => {
+    controller.restoreNetwork(); // causes page reload, no need to set state
+  };
+
+  // Share/Download functions
+  const handleCopyLink = async () => {
+    handleMenuClose();
+    await handleCopyToClipboard(); 
+    snack.showMessage("Link copied to clipboard");
+  };
+
+  const handleExportImages = async () => {
+    handleMenuClose();
+    setImageExportEnabled(false);
+    snack.showSpinner("Preparing network images...");
+    await delay(50); // allows the menu to close immediately, otherwise it hangs for a couple seconds
+    await controller.exportImageArchive(controller);
+    snack.close();
+    setImageExportEnabled(true); 
+  };
+
+  const handleExportData = async () => {
+    handleMenuClose();
+    setDataExportEnabled(false);
+    const promise = controller.exportDataArchive(controller);
+    await spinnerUntilDone(promise, "Preparing enrichment data...");
+    setDataExportEnabled(true);
+  };
+
+  const spinnerUntilDone = async (promise, message) => {
+    // Only show spinner if it takes longer than the delay
+    const value = await Promise.race([ promise, delay(500) ]);
+    if (value === 'delay') {
+      // if the delay promise finished first
+      snack.showSpinner(message);
+    }
+    await promise; // wait for the export to finish if it hasn't already
+    snack.close();
+  };
+
+  // Definitons for the toolbar (or mobile menu drawer)
+  const shareMenuDef = [
+    {
+      title: "Share Link to Network",
+      icon: <LinkIcon />,
+      onClick: handleCopyLink,
+    }, {
+      title: "Download Network Images",
+      icon: <InsertDriveFileOutlinedIcon />,
+      disabled: !imageExportEnabled,
+      onClick: handleExportImages,
+    }, {
+      title: "Download Enrichment Data",
+      icon: <InsertDriveFileOutlinedIcon />,
+      disabled: !dataExportEnabled,
+      onClick: handleExportData,
+    }
+  ];
+
+  const menuDef = [ 
+    {
+      title: getUndoMenuTitle(undoType),
+      icon: <UndoIcon />,
+      onClick: () => controller.undoHandler.undo(),
+      isEnabled: () => undoEnabled,
+    },
+    {
+      title: "Delete Selected Nodes",
+      icon: <DeleteIcon />,
+      onClick: () => controller.deleteSelectedNodes(),
+    }, {
+      title: "Restore Network to Initial Layout",
+      icon: <RestoreIcon />,
+      onClick: handleNetworkRestore,
+      unrelated: true,
+    }, {
+      title: "Zoom In",
+      icon: <AddIcon />,
+      onClick: panner.zoomIn,
+    }, {
+      title: "Zoom Out",
+      icon: <RemoveIcon />,
+      onClick: panner.zoomOut,
+    }, {
+      title: "Fit Figure to Screen",
+      icon: <FitScreenIcon />,
+      onClick: panner.fit,
+      unrelated: true,
+    }, {
+      title: "Share/Download",
+      icon: <CloudDownloadIcon />,
+      onClick: showShareMenu,
+      subMenu: shareMenuDef,
+    },
+  ];
+
+  useEffect(() => {
+    const onUndo = (pushOrPop, type, empty, peekType) => {
+      setUndoEnabled(!empty);
+      setUndoType(peekType);
+    };
+    controller.bus.on('undo', onUndo);
+    return () => controller.bus.removeListener('undo', onUndo);
+  }, []);
+
+  useEffect(() => {
+    Mousetrap
+      .bind(['-','_'], panner.zoomOut)
+      .bind(['=','+'], panner.zoomIn)
+      .bind('up', panner.panUp)
+      .bind('down', panner.panDown)
+      .bind('left', panner.panLeft)
+      .bind('right', panner.panRight)
+      .bind(['f', 'space'], panner.fit) 
+      .bind(['backspace','del'], () => controller.deleteSelectedNodes());
+  
+    return () => Mousetrap.unbind(['-','_','=','+','up','down','left','right','f','space','backspace','del']);
+  }, [panner]);
+
   return (
-    <div className={classes.root} onClick={onContentClick}>
-      <LeftDrawer
-        open={openLeftDrawer}
-        isMobile={isMobile}
+    <>
+      <Header
+        menuDef={menuDef}
         controller={controller}
-        onClose={onCloseLeftDrawer}
-      />
-      <div className={classes.background}>
-        <div
-          className={clsx(classes.cy, { [classes.cyShiftX]: shiftXCy })}
-          style={shiftYCy ? {height: `calc(100% - ${bottomDrawerHeight()}px)`,} : {}}
-        >
-          <div id="cy" className={classes.cy} style={{ zIndex: 1, width: '100%', height: '100%' }} />
-          <NetworkBackground controller={controller} />
-        </div>
-      </div>
-      <BottomDrawer
-        open={openBottomDrawer}
-        isMobile={isMobile}
         leftDrawerOpen={openLeftDrawer}
-        onToggle={onToggleBottomDrawer}
-        controller={controller}
+        isMobile={isMobile}
+        onOpenLeftDrawer={onOpenLeftDrawer}
+        onOpenRightDrawer={onOpenRightDrawer}
       />
-    </div>
+      <div className={classes.root} onClick={onContentClick}>
+        <LeftDrawer
+          open={openLeftDrawer}
+          isMobile={isMobile}
+          controller={controller}
+          onClose={onCloseLeftDrawer}
+        />
+        <div className={classes.background}>
+          <div
+            className={clsx(classes.cy, { [classes.cyShiftX]: shiftXCy })}
+            style={shiftYCy ? {height: `calc(100% - ${bottomDrawerHeight()}px)`,} : {}}
+          >
+            <div id="cy" className={classes.cy} style={{ zIndex: 1, width: '100%', height: '100%' }} />
+            <NetworkBackground controller={controller} />
+          </div>
+          <RightDrawer
+            open={openRightDrawer}
+            menu={menuDef}
+            onClose={onCloseRightDrawer}
+          />
+        </div>
+        <BottomDrawer
+          open={openBottomDrawer}
+          isMobile={isMobile}
+          leftDrawerOpen={openLeftDrawer}
+          onToggle={onToggleBottomDrawer}
+          controller={controller}
+        />
+      </div>
+    {!isMobile && anchorEl && (
+      <PopoverMenu
+        open={menuName === SHARE_MENU_ID}
+        target={anchorEl}
+        menu={shareMenuDef}
+        onClose={handleMenuClose}
+      />
+    )}
+      <RestoreConfirmDialog 
+        open={confirmDialogOpen} 
+        onOk={onConfirmOk} 
+        onCancel={onConfirmCancel} 
+      />
+      <Snackbar
+        className={classes.snackBar}
+        anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+        open={snackBarState.open || false} 
+        autoHideDuration={snackBarState.autoHideDelay || null} 
+        onClose={() => setSnackBarState({ open: false })} 
+      >
+        <SnackbarContent 
+          className={classes.snackBarContent}
+          message={<span>{snackBarState.message || ""}</span>}
+          action={(() => {
+            if(snackBarState.closeable) {
+              return (
+                <IconButton size='small' onClick={() => setSnackBarState({ open: false })}>
+                  <CloseIcon />
+                </IconButton>
+              );
+            } else if(snackBarState.spinner) {
+              return <CircularProgressIcon size={20}/>;
+            }
+          })()}
+        />
+      </Snackbar>
+    </>
   );
 };
-
-NetworkBackground.propTypes = {
-  controller: PropTypes.instanceOf(NetworkEditorController).isRequired,
-};
 Main.propTypes = {
-  controller: PropTypes.instanceOf(NetworkEditorController),
+  controller: PropTypes.instanceOf(NetworkEditorController).isRequired,
   openLeftDrawer: PropTypes.bool.isRequired,
+  openRightDrawer: PropTypes.bool.isRequired,
   openBottomDrawer: PropTypes.bool.isRequired,
   isMobile: PropTypes.bool.isRequired,
   onContentClick: PropTypes.func.isRequired,
   onCloseLeftDrawer: PropTypes.func.isRequired,
+  onCloseRightDrawer: PropTypes.func.isRequired,
+  onOpenLeftDrawer: PropTypes.func.isRequired,
+  onOpenRightDrawer: PropTypes.func.isRequired,
   onToggleBottomDrawer: PropTypes.func,
 };
