+++ src/client/components/network-editor/util.js
@@ -1,0 +1,8 @@
+
+export function delay(millis) {
+  return new Promise(r => setTimeout(r, millis, 'delay'));
+}
+
+export function stringToBlob(str) {
+  return new Blob([str], { type: 'text/plain;charset=utf-8' });
+}
